# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TALABAT Multi-Restaurant SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all restaurants
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantListResponse"

    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create a new restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found

    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"

    delete:
      operationId: deleteRestaurant
      tags: [restaurants]
      summary: Delete restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
  /restaurants/{id}/status:
    patch:
      operationId: updateRestaurantStatus
      tags: [restaurants]
      summary: Activate or suspend restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"

  # ── Subscriptions ───────────────────────────────────────────────────────────
  /restaurants/{id}/subscription:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get restaurant subscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create or renew subscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Extend or change subscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListResponse"

  # ── Categories ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Restaurant:
      type: object
      required: [id, name, slug, phone, address, status, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        slug: { type: string }
        phone: { type: string }
        address: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        primaryColor: { type: ["string", "null"] }
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
        whatsappNumber: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        subscription:
          $ref: "#/components/schemas/Subscription"
    RestaurantInput:
      type: object
      required: [name, slug, phone, address, subscriptionPlan]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1, pattern: "^[a-z0-9-]+$" }
        phone: { type: string }
        address: { type: string }
        logoUrl: { type: string }
        coverUrl: { type: string }
        description: { type: string }
        primaryColor: { type: string }
        whatsappNumber: { type: string }
        subscriptionPlan:
          type: string
          enum: [TRIAL, MONTHLY, YEARLY]
    RestaurantPatch:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        primaryColor: { type: ["string", "null"] }
        whatsappNumber: { type: ["string", "null"] }
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
    RestaurantListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Restaurant"
        total: { type: number }
        page: { type: number }
        limit: { type: number }

    # ── Subscription ────────────────────────────────────────────────────────────
    Subscription:
      type: object
      required: [id, restaurantId, plan, status, startDate, expiryDate]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        plan:
          type: string
          enum: [TRIAL, MONTHLY, YEARLY]
        status:
          type: string
          enum: [TRIAL, ACTIVE, EXPIRED, SUSPENDED, INACTIVE]
        startDate: { type: string, format: date }
        expiryDate: { type: string, format: date }
        createdAt: { type: string, format: date-time }
    SubscriptionInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [TRIAL, MONTHLY, YEARLY]
    SubscriptionPatch:
      type: object
      properties:
        plan:
          type: string
          enum: [TRIAL, MONTHLY, YEARLY]
        status:
          type: string
          enum: [TRIAL, ACTIVE, EXPIRED, SUSPENDED, INACTIVE]
        extensionDays: { type: number }
    SubscriptionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        total: { type: number }
        page: { type: number }
        limit: { type: number }

    # ── Category ────────────────────────────────────────────────────────────────
...